# perf(printing): make print_seq O(n) rather than quadratic