@oscarbenjamin 2021-03-25
With the PR:
```python
In [3]: xs = symbols('x:1000')

In [4]: eqs = [x1 + sqrt(2)*x2 for x1, x2 in zip(xs[:-1], xs[1:])]

In [5]: eqs.append(xs[-1] - sqrt(3))

In [6]: s = linsolve(eqs, xs)

In [7]: %time ok = pretty(s)
CPU times: user 10.6 s, sys: 174 ms, total: 10.8 s
Wall time: 11 s
```
On master we have:
```python
In [5]: %time ok = pretty(s)
CPU times: user 29.9 s, sys: 66.6 ms, total: 30 s
Wall time: 30 s
```
So that's quite a bit faster. It removes a quadratic algorithm and replaces it with one that is more linear. There might be other quadratic algorithms in there because it is still too slow.